Rename class 'JunitExporter' to 'TestSuites'
