+++ common/junit_exporter/include/junit_exporter/test_suites.hpp
@@ -12,18 +12,69 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef JUNIT_EXPORTER__TEST_SUITES_HPP_
-#define JUNIT_EXPORTER__TEST_SUITES_HPP_
+#ifndef JUNIT_EXPORTER__JUNIT_EXPORTER_HPP_
+#define JUNIT_EXPORTER__JUNIT_EXPORTER_HPP_
 
+#include <boost/filesystem/path.hpp>
 #include <junit_exporter/test_suite.hpp>
-#include <unordered_map>
+#include <numeric>
+#include <pugixml.hpp>
+#include <string>
 
 namespace common
 {
 inline namespace junit
 {
-using TestSuites = std::unordered_map<std::string, TestSuite>;
+class TestSuites
+{
+  const std::string timestamp_;
+
+  std::unordered_map<std::string, TestSuite> test_suites_;
+
+public:
+  explicit TestSuites();
+
+  void write(const boost::filesystem::path & path);
+
+  auto contains(const std::string & suite_name, const std::string & case_name) const
+  {
+    const auto iter = test_suites_.find(suite_name);
+
+    if (iter != std::end(test_suites_)) {
+      return std::any_of(
+        std::begin(iter->second), std::end(iter->second),
+        [&](const auto & each_case) { return each_case.name == case_name; });
+    } else {
+      return false;
+    }
+  }
+
+  auto getTotalTime() const
+  {
+    return std::accumulate(
+      std::cbegin(test_suites_), std::cend(test_suites_), 0,
+      [](const auto sum, const auto & each_suite) {
+        const auto & each_cases = std::get<1>(each_suite);
+
+        return sum + std::accumulate(
+                       std::cbegin(each_cases), std::cend(each_cases), 0,
+                       [](const auto sum, const auto & each_case) { return sum + each_case.time; });
+      });
+  }
+
+  template <typename... Ts>
+  void addTestCase(
+    const std::string & suite_name,  //
+    const std::string & case_name,   //
+    Ts &&... xs)
+  {
+    if (not contains(suite_name, case_name)) {
+      test_suites_[suite_name].emplace_back(
+        case_name, suite_name, std::forward<decltype(xs)>(xs)...);
+    }
+  }
+};
 }  // namespace junit
 }  // namespace common
 
-#endif  // JUNIT_EXPORTER__TEST_SUITES_HPP_
+#endif  // JUNIT_EXPORTER__JUNIT_EXPORTER_HPP_
